+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-03-18
+
+### Added
+
+- Added new pipeline method `add_tasks_to_workflow` for adding tasks to an existing Jobmon workflow
+- Added new Jobmon-related arguments: `external_upstream_dependencies`, `task_and_template_prefix`, and `max_attempts`.
+
 ## [1.0.3] - 2025-03-12
 
 ### Changed